+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### Unreleased
+
+* Breaking changes
+  * Parsing fails more eagerly. This affects lazy list parsing and parsing via
+    the `Regex.Base` functions `prepareParser` and `stepParser`.
+
 ### 0.1.0.0 -- 2024-03-04
 
 * First version.